backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
requests
python-dotenv
polyline
gunicorn
whitenoise
